webapi/webapi-webrtc-w3c-tests/webrtc/RTCDataChannel_onbufferedamountlow.html

<!DOCTYPE html>
<!--
Copyright (c) 2015 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-->

<meta charset="utf-8">
<meta name="timeout" content="long">
<title>RTCDataChannel onbufferedamountlow event</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Chunyan Wang" href="mailto:chunyanx.wang@intel.com">
<link rel="help" href="http://w3c.github.io/webrtc-pc/#widl-RTCDataChannel-onbufferedamountlow">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="support/support.js"></script>
<div id="log"></div>
<script>
async_test(function(t) {
  var localConnection = new RTCPeerConnection(null);
  var remoteConnection = new RTCPeerConnection(null);
  var sendChannel = localConnection.createDataChannel("sendDataChannel");
  var receiveChannel;
  
  assert_equals(sendChannel.bufferedAmountLowThreshold, 0);
  assert_equals(typeof sendChannel.bufferedAmountLowThreshold, "number");

  var size = 16384;
  var buffer = new ArrayBuffer(size);
  sendChannel.bufferedAmountLowThreshold = size / 2;

  sendChannel.onopen = t.step_func(function(evt) {
    if(sendChannel.readyState == "open") {
      sendChannel.addEventListener('bufferedamountlow', t.step_func(function(evt) {
        assert_less_than_equal(sendChannel.bufferedAmount, sendChannel.bufferedAmountLowThreshold);
        t.done();
      }));
      for(var i = 0; i < 100; i++) {
        sendChannel.send(buffer);
      }
    }
  });

  localConnection.onicecandidate = t.step_func(function(evt) {
    if(evt.candidate) {
      remoteConnection.addIceCandidate(event.candidate);
    }
  });

  remoteConnection.onicecandidate = t.step_func(function(evt) {
    if(evt.candidate) {
      localConnection.addIceCandidate(event.candidate);
    }
  });

  localConnection.createOffer(function getDescription(desc) {
    t.step(function() {
      localConnection.setLocalDescription(desc);
      remoteConnection.setRemoteDescription(desc);
      remoteConnection.createAnswer(function getAnswer(answer) {
        t.step(function() {
          remoteConnection.setLocalDescription(answer);
          localConnection.setRemoteDescription(answer);
        });
      }, function showAnswerError(error) {
        t.step(function() {
          assert_unreached(error.name + ":" + error.message);
        });
      });
    });
  }, function showConnectionError(error) {
    t.step(function() {
      assert_unreached(error.name + ":" + error.message);
    });
  });
}, "Check that RTCDataChannel monotors with bufferedamountlow event.");

</script>
